+++ src/lib/timeseries.js
@@ -241,3 +241,27 @@ export const sortTimeseriesByCO2GroupedByPurpose = aggregatedTimeseries => {
 
   return sortGroupedTimeseries(timeseriesByPurposes, OTHER_PURPOSE)
 }
+
+export const getStartDate = timeserie => {
+  return new Date(timeserie.startDate)
+}
+
+export const getEndDate = timeserie => {
+  return new Date(timeserie.endDate)
+}
+
+export const getStartPlaceDisplayName = timeserie => {
+  return get(timeserie, 'series[0].features[0].properties.display_name')
+}
+
+export const getEndPlaceDisplayName = timeserie => {
+  return get(timeserie, 'series[0].features[1].properties.display_name')
+}
+
+export const getGeoJSONData = timeserie => {
+  return get(timeserie, 'series[0]')
+}
+
+export const getManualPurpose = timeserie => {
+  return get(timeserie, 'series[0].properties.manual_purpose')
+}